:truck: Rename LSP `basic_structures` -> `common_structures`
